+++ nnetsauce/predictioninterval/predictioninterval.py
@@ -140,17 +140,6 @@ class PredictionInterval(BaseEstimator, RegressorMixin):
             self.icp_ = IcpRegressor(nc)
             self.icp_.fit(X_train, y_train)
             self.icp_.calibrate(X_calibration, y_calibration)
-            preds_calibration = self.icp_.predict(X_calibration)
-            self.calibrated_residuals_ = y_calibration - preds_calibration
-            absolute_residuals = np.abs(self.calibrated_residuals_)
-            self.calibrated_residuals_scaler_ = StandardScaler(
-                with_mean=True, with_std=True
-            )
-            self.scaled_calibrated_residuals_ = (
-                self.calibrated_residuals_scaler_.fit_transform(
-                    self.calibrated_residuals_.reshape(-1, 1)
-                ).ravel()
-            )
 
         return self